voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,mcp]~=1.2.18",
    "python-dotenv>=1.1.1",
    "websockets>=12.0",
    "e2b-desktop>=2.1.0",
    "httpx>=0.27.0",
    "mem0ai>=1.0.1",
    "psycopg2-binary>=2.9.11",
]
